feat(dsg) fix code review comments
